Fix cluster / io_ctx resource / memory leaks
When the user forgot to call the destroy/shutdown functions,
phprados was leaking memory since the instances never get
cleaned up. This can be an issue especially when executing
in the scope of a long-running process (like as Apache module).

Also introduce a ioctx counter in php_rados_cluster to prevent
shutting down with active io contexts.